fix bugs

+++ models/english/RMS/utils.py
@@ -21,6 +21,6 @@ paths = Paths()
 def rms(iterator, language, parameters):
     frame_rate = parameters['frame_rate']
 
-    return np.apply_along_axis(lambda y: np.sqrt(np.mean(np.square(y))),1, iterator)
+    return np.apply_along_axis(lambda y: np.sqrt(np.mean(np.square(y, dtype=np.float64))),1, iterator)
 
     #return np.apply_along_axis(lambda y: simps(np.abs(hilbert(y)), dx=1/frame_rate),1, iterator)
